src/main/java/com/bod/entity/FoodBuilder.java
package com.bod.entity;

public class FoodBuilder {
    private Food foodUnit;

    public void reset() {
        foodUnit = new Food();
    }

    public void setName(String name) {
        foodUnit.setName(name);
    }

    public void setAmount(double amount) {
        foodUnit.setAmount(amount);
    }

    public void setNumber(int number) {
        foodUnit.setNumber(number);
    }

    public void setCalories(double calories) {
        foodUnit.setCalories(calories);
    }

    public void setProteins(double proteins) {
        foodUnit.setProtein(proteins);
    }

    public void setFats(double fats) {
        foodUnit.setFat(fats);
    }

    public void setCarbohydrates(double carbohydrates) {
        foodUnit.setCarbohydrates(carbohydrates);
    }

    public Food getResult() {
        return foodUnit;
    }
}
